+++ renderer/__init__.py
@@ -1,5 +1,4 @@
 from renderer.base import *
-import renderer.tools
 import renderer.mathtools
 import renderer.validate
 import renderer.misc