+++ renderer/tools/nodes.py
@@ -1,0 +1,33 @@
+import blessed
+
+import renderer.internals.internal as internal # type: ignore
+import renderer.validate as validate
+from renderer.types import *
+
+term = blessed.Terminal()
+
+def find_components_attached(node_id: str, component_json: ComponentJson) -> List[Tuple[str, int]]:
+    """
+    Finds which components attach to a node.
+    More info: https://tile-renderer.readthedocs.io/en/latest/functions.html#renderer.tools.nodes.findPLAsAttached
+    """
+    components = []
+    for component_id, component in component_json.items():
+        #print(component_id)
+        if node_id in component['nodes']:
+            components.append((component_id, component['nodes'].index(node_id)))
+    return components
+
+def to_coords(nodes: List[str], node_json: NodeJson) -> List[Coord]:
+    """
+    Converts a list of nodes IDs into a list of coordinates with a node dictionary/JSON as its reference.
+    More info: https://tile-renderer.readthedocs.io/en/latest/functions.html#renderer.tools.nodes.toCoords
+    """
+    validate.v_node_json(node_json)
+    validate.v_node_list(nodes, node_json)
+    coords = []
+    for node_id in nodes:
+        if node_id not in node_json.keys():
+            raise KeyError(f"Node '{node_id}' does not exist")
+        coords.append((node_json[node_id]['x'], node_json[node_id]['y']))
+    return coords